feat: Add default, low prio noopSchemaStore into AutoConfiguration
