+++ horizon-spring-boot-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,3 +1,3 @@
 de.telekom.eni.pandora.horizon.autoconfigure.HorizonAutoConfiguration
 de.telekom.eni.pandora.horizon.autoconfigure.KubernetesClientConfiguration
-de.telekom.eni.pandora.horizon.autoconfigure.cache.SchemaCacheAutoConfiguration
+de.telekom.eni.pandora.horizon.autoconfigure.schema.SchemaStoreConfiguration
